// liars-bar-backend/src/game/action/event.enum.ts
export enum GameEventType {
  GAME_STARTED = 'GAME_STARTED',
  GAME_ENDED = 'GAME_ENDED',
  PLAYER_INITIALIZED = 'PLAYER_INITIALIZED',
  ROUND_STARTED = 'ROUND_STARTED',
  ROUND_ENDED = 'ROUND_ENDED',
  CLAIM_MADE = 'CLAIM_MADE',
  CHALLENGE_MADE = 'CHALLENGE_MADE',
  CHALLENGE_RESULT = 'CHALLENGE_RESULT',
  PLAYER_SHOT = 'PLAYER_SHOT',
  PLAYER_ELIMINATED = 'PLAYER_ELIMINATED',
  TURN_CHANGED = 'TURN_CHANGED',
  ROOM_JOINED = 'ROOM_JOINED',
}